// types/product.ts
export interface Product {
  id: number;
  title: string;
  description?: string;
  price: number;
  discountPercentage?: number;
  rating?: number;
  stock?: number;
  brand?: string;
  category?: string;
  thumbnail: string;
  images: string[]; 
  tags?: string[];
}

export interface RouteParams {
  id: string;
}